feat: ChannelManager에 Header 컴포넌트 추가
- ChannelManager에서도 Header 컴포넌트에 존재하는 기능을 사용하기 위해서
- Body의 내용을 DashBoard 컴포넌트로 분리함.
