+++ frontend/src/components/DashBoard/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './DashBoard';
